Use single controller for both actions
